More linux issues

+++ picoquicfirst/picoquicdemo.c
@@ -420,7 +420,7 @@ int quic_client(const char* ip_address_text, int server_port,
     struct sockaddr_storage client_address;
     struct sockaddr_storage packet_from;
     struct sockaddr_storage packet_to;
-    int if_index_to;
+    unsigned long if_index_to;
     uint8_t buffer[1536];
     uint8_t send_buffer[1536];
     size_t send_length = 0;
